From: Torbjörn Svensson
As I told you on IRC, I've found a bug in icecream related to .d  
files. If you have -MD (or -MMD) and icecream should distribute the  
job, then the .d will be creaded in the source tree and that's not  
always what you want. My patch puts the .d file in the same directory  
as the GCC would have done if invoked. The trick I'm using is to put  
the -MF option in the local flags so that the .d file will be placed  
in the target tree while the pre-processed C/C++ file is still dumped  
on stdout.


git-svn-id: svn://anonsvn.kde.org/home/kde/trunk/icecream@1222062 283d02a7-25f6-0310-bc7c-ecb5cbfe19da